Refactored common functionality for saving geo hazards and days back
