+++ src/app/modules/common-core/helpers/arrays.ts
@@ -1,0 +1,3 @@
+export function isArraysEqual(array1: number[], array2: number[]): boolean {
+  return array1.length === array2.length && array1.every((value, index) => value === array2[index]);
+}
