+++ src/app/modules/side-menu/components/observations-days-back/observations-days-back.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
-import { Subscription, firstValueFrom } from 'rxjs';
+import { Subscription } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { UserSettingService } from '../../../../core/services/user-setting/user-setting.service';
 import { GeoHazard } from 'src/app/modules/common-core/models';
@@ -46,18 +46,8 @@ export class ObservationsDaysBackComponent extends NgDestoryBase implements OnIn
   }
 
   async save(): Promise<void> {
-    const userSetting = await firstValueFrom(this.userSettingService.userSetting$);
-    let changed = false;
-    for (const geoHazard of userSetting.currentGeoHazard) {
-      const existingValue = userSetting.observationDaysBack.find((x) => x.geoHazard === geoHazard);
-      if (existingValue.daysBack !== this.selectedDaysBack) {
-        existingValue.daysBack = this.selectedDaysBack;
-        changed = true;
-      }
-    }
-    if (changed) {
-      this.userSettingService.saveUserSettings(userSetting);
-      const daysBack = await firstValueFrom(this.userSettingService.daysBackForCurrentGeoHazard$);
+    const daysBack = await this.userSettingService.saveGeoHazardsAndDaysBack({ daysBack: this.selectedDaysBack });
+    if (typeof daysBack === 'number') {
       this.changeDaysBack.emit(daysBack);
     }
   }
